Added health fields
Edited some comments. I also implemented health fields. A deer now needs to have a health of at least 5 in addition to being the appropriate age in order to breed.

I added a statement which decrements health if the deer is unable to find food. Health also increments if the deer is able to find food.